:art:: Rename & cleanup utils files
