+++ src/components/application-map/ApplicationMap.tsx
@@ -6,7 +6,7 @@ import { useEffect } from 'react';
 
 import { RankDirectionType, isApplication } from '../../types';
 import { ArgoApplication, ArgoApplicationSet } from '../../types/argocd';
-import { applyLayoutToGraph } from '../../utils/graph';
+import { applyLayoutToGraph } from '../../utils/applyLayoutToGraph';
 import { ApplicationMapNavigationControls } from './navigation-controls';
 import { ApplicationMapNode } from './node';
 import type { ApplicationMapNodeType } from './node';
