+++ src/utils/hasCycle.ts
