+++ src/utils/resourceId.ts
